Factor out `ChainDBTrace` & `SubscriptionTrace` to their own modules
