+++ cardano-node/cardano-node.cabal
@@ -44,6 +44,8 @@ library
 
 
   other-modules:       Paths_cardano_node
+                       Cardano.Tracing.ToObjectOrphans.ChainDBTrace
+                       Cardano.Tracing.ToObjectOrphans.SubscriptionTrace
                        Cardano.Tracing.ToObjectOrphans
                        Cardano.Tracing.MicroBenchmarking
